Allow for configuration file to be merged with defaults
The config_file method from Sinatra::ConfigFile can take multiple
path parameters. With this in mind the default configuration path
is now the first parameter and the user defined configuration path
is the second argument. When loaded, if the second file contains a
key that was already in the first file, that key will be overwritten
with the value of the key from the second file. If the second file
does not exist, then only the first file will be loaded into the
settings.