Ensure that loglevel is using the const value
When loglevel is set in server.yml, it is passed in as a String, but
WEBrick::Log @loglevel requires an Integer. To resolve this, when
LOGGER is set, it will now grab the value of the constant that is
associated with the loglevel and pass that value in (which is an integer).