Fix a http body param bug
Fixed a bug where one of the parameters sent to a method was named
`body`. This was an issue as the `body` method is built into Rack and
was causing the application to become confused are fail on API calls.

Updated and organized .gitignore.

Gemfile.lock  updated as part of a bundle install